AP_InertialSensor: fix -Werror=sign-compare in QMI8658A/HXY42688 vendor drivers
The register-check loops compared a signed `int` index against
`sizeof(COMM_REGS)/sizeof(COMM_REGS[0])` (size_t), which the SITL/Linux CI
toolchain rejects under -Werror=sign-compare (build (base)/(clang) jobs failed
at "build ... base"; -Wfatal-errors masked the second driver). The ARM gcc-10
board build did not flag it, so the four -main board binaries are unaffected
(loop behavior is identical) -- this is a compile-cleanliness fix only.

Change the loop index to size_t in both regs-check loops. Validated: SITL copter
build compiles both drivers clean.

+++ libraries/AP_InertialSensor/AP_InertialSensor_QMI8658A.cpp
@@ -167,7 +167,7 @@ bool AP_InertialSensor_QMI8658A::init()
 
 bool AP_InertialSensor_QMI8658A::regs_check_out()
 {
-    for (int i = 0; i < sizeof(AP_InertialSensor_QMI8658A::COMM_REGS)/sizeof(AP_InertialSensor_QMI8658A::COMM_REGS[0]); ++i) {
+    for (size_t i = 0; i < sizeof(AP_InertialSensor_QMI8658A::COMM_REGS)/sizeof(AP_InertialSensor_QMI8658A::COMM_REGS[0]); ++i) {
         uint8_t reg = AP_InertialSensor_QMI8658A::COMM_REGS[i][0];
         uint8_t value = read_register(reg);
         if (value != AP_InertialSensor_QMI8658A::COMM_REGS[i][1]) {
